Added code editor
The code editor allows you to input code, then eval it with a warning.
The textarea has functionality for adding tabs without changing the focused element, as well as doing ctrl+d to duplicate the currently selected lines.
The output of the exec is shown below the text area. It consists of all the calls to output(str) concatenated, in red if there is an error, and the return value of the eval in its own span.